@@ -1,15 +1,15 @@
 {
   "name": "dynamic-pie-chart",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "peerDependencies": {
-    "@angular/common": "^9.1.11",
-    "@angular/core": "^9.1.11",
-    "@angular/platform-browser": "^9.1.11",
-    "@angular/platform-browser-dynamic": "^9.1.11",
+    "@angular/common": "~10.2.5",
+    "@angular/core": "~10.2.5",
+    "@angular/platform-browser": "~10.2.5",
+    "@angular/platform-browser-dynamic": "~10.2.5",
     "d3": "^5.16.0"
   },
   "dependencies": {
-    "tslib": "^1.10.0"
+    "tslib": "^2.0.0"
   },
   "license": "MIT",
   "repository": {
